+++ dwds/lib/dwds.dart
@@ -90,6 +90,7 @@ class Dwds {
     bool verbose,
     UrlEncoder urlEncoder,
     bool useFileProvider = false,
+    bool spawnDds = true,
     // TODO(annagrin): make expressionCompiler argument required
     // [issue 881](https://github.com/dart-lang/webdev/issues/881)
     ExpressionCompiler expressionCompiler,
@@ -150,6 +151,7 @@ class Dwds {
       serveDevTools,
       expressionCompiler,
       injected,
+      spawnDds,
     );
 
     return Dwds._(