+++ webdev/lib/src/command/serve_command.dart
@@ -44,6 +44,11 @@ refresh: Performs a full page refresh.
         help: 'Specify which port the Chrome debugger is listening on. '
             'If used with $launchInChromeFlag Chrome will be started with the'
             ' debugger listening on this port.')
+    ..addFlag(disableDdsFlag,
+        negatable: false,
+        help: 'Disable the Dart Development Service (DDS). Disabling DDS may '
+        'result in a degraded developer experience in some tools.',
+        hide: true)
     ..addOption(hostnameFlag,
         help: 'Specify the hostname to serve on.', defaultsTo: 'localhost')
     ..addFlag(hotRestartFlag,